C code completed

+++ include/raytracing.h
@@ -5,6 +5,6 @@
 #include "sphere.h"
 #include "camera.h"
 
-void raytracing(color_t* image, const uint16_t width, const uint16_t height, const uint16_t raysPerPixel, const uint8_t raysDepth, const sphere_t* spheres, const uint8_t numberOfSpheres, const camera_t* camera);
+void raytracing(color_t* image, const uint16_t width, const uint16_t height, const uint16_t raysPerPixel, const uint8_t raysDepth, const sphere_t* spheres, const uint16_t numberOfSpheres, const camera_t* camera);
 
 #endif
